+++ praw/models/reddit/mixins/__init__.py
@@ -2,6 +2,7 @@
 from json import dumps
 
 from ....const import API_PATH
+from ...util import stream_generator
 from .editable import EditableMixin
 from .fullname import FullnameMixin
 from .gildable import GildableMixin
@@ -294,3 +295,34 @@ class UserContentMixin(
     VotableMixin,
 ):
     """A convenience mixin that applies to both Comments and Submissions."""
+
+    def comment_stream(self, **stream_options):
+        """Create a stream yielding new top-level comments/comment replies.
+
+        :param stream_options: Options to pass to :func:`.stream_generator`.
+        :returns: A stream for top-level comments.
+
+        Example usage:
+
+        .. code-block:: python
+
+            submission = reddit.submission(id='5or86n')
+            for comment in submission.comment_stream():
+                print(comment)
+
+
+            parent_comment = reddit.comment(id='dxolpyc')
+            for comment in parent_comment.comment_stream():
+                print(comment)
+        """
+        forest = self.comments if "comments" in dir(self) else self.replies
+        return stream_generator(
+            forest.tree,
+            exclude_before=True,
+            showmore=False,
+            sort="new",
+            context=0,
+            depth=0,
+            threaded=False,
+            **stream_options
+        )
